+++ android/contrib/compile-ffmpeg.sh
@@ -26,7 +26,7 @@ set -e
 set +x
 
 FF_ACT_ARCHS_32="armv5 armv7a x86"
-FF_ACT_ARCHS_64="armv5 armv7a x86 arm64"
+FF_ACT_ARCHS_64="armv5 armv7a arm64 x86 x86_64"
 FF_ACT_ARCHS_ALL=$FF_ACT_ARCHS_64
 
 echo_archs() {
@@ -40,7 +40,7 @@ echo_archs() {
 
 echo_usage() {
     echo "Usage:"
-    echo "  compile-ffmpeg.sh armv5|armv7a|x86|arm64"
+    echo "  compile-ffmpeg.sh armv5|armv7a|arm64|x86|x86_64"
     echo "  compile-ffmpeg.sh all|all32"
     echo "  compile-ffmpeg.sh all64"
     echo "  compile-ffmpeg.sh clean"
@@ -63,7 +63,7 @@ case "$FF_TARGET" in
         echo_archs armv7a
         sh tools/do-compile-ffmpeg.sh armv7a
     ;;
-    armv5|armv7a|x86|arm64)
+    armv5|armv7a|arm64|x86|x86_64)
         echo_archs $FF_TARGET
         sh tools/do-compile-ffmpeg.sh $FF_TARGET
         echo_nextstep_help