+++ ijkmedia/ijkplayer/ff_ffplay_options.h
@@ -86,6 +86,9 @@ static const AVOption ffp_context_options[] = {
         OPTION_OFFSET(overlay_format),  OPTION_INT(SDL_FCC_RV32, INT_MIN, INT_MAX),
         .unit = "overlay-format" },
     { "fcc-i420",                       "", 0, OPTION_CONST(SDL_FCC_I420), .unit = "overlay-format" },
+#ifdef __APPLE__
+    { "fcc-i4al",                       "", 0, OPTION_CONST(SDL_FCC_I444P10LE), .unit = "overlay-format" },
+#endif
     { "fcc-yv12",                       "", 0, OPTION_CONST(SDL_FCC_YV12), .unit = "overlay-format" },
     { "fcc-rv16",                       "", 0, OPTION_CONST(SDL_FCC_RV16), .unit = "overlay-format" },
     { "fcc-rv24",                       "", 0, OPTION_CONST(SDL_FCC_RV24), .unit = "overlay-format" },