+++ ios/IJKMediaDemo/IJKMediaDemo/IJKMoviePlayerViewController.h
@@ -29,7 +29,7 @@
 
 - (IBAction)onClickMediaControl:(id)sender;
 - (IBAction)onClickOverlay:(id)sender;
-- (IBAction)onClickBack:(id)sender;
+- (IBAction)onClickDone:(id)sender;
 - (IBAction)onClickPlay:(id)sender;
 - (IBAction)onClickPause:(id)sender;