+++ NEWS.md
@@ -3,6 +3,7 @@ tag next
 - ios: support playbackRate change. (iOS 7.0 or later)
 - android: support speed change. (Android 6.0 or later)
 - player: do not link avfilter by default.
+- android: add x86_64 support
 
 tag k0.4.4
 --------------------------------